services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
asyncpg==0.29.0
aiosqlite==0.19.0
pulsar-client==3.4.0
aiokafka==0.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
-e ../../packages/schemas
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
python-dotenv==1.0.0